Génération d'un mouvement vide pour les drms à néants

+++ project/test/unit/giilda54ExportDRMCIVADB2Test.php
@@ -20,7 +20,9 @@ foreach(DRMClient::getInstance()->viewByIdentifiant($viti->identifiant) as $k =>
   $drm = DRMClient::getInstance()->find($k);
   $drm->delete(false);
   $csv = CSVDRMClient::getInstance()->find(str_replace("DRM", "CSVDRM", $k));
-  $csv->delete(false);
+    if($csv) {
+        $csv->delete(false);
+    }
 }
 
 $preLigneCaveCSV = "CAVE;".$periode.";".$viti->identifiant.";";
@@ -74,7 +76,9 @@ $temp = fopen($tmpfname, "w");
 fwrite($temp, $csv);
 fclose($temp);
 
-$t = new lime_test(18);
+$t = new lime_test(32);
+
+$t->comment("Export d'une DRM");
 
 $drm = DRMClient::getInstance()->createDoc($viti->identifiant, $periode);
 $import = new DRMImportCsvEdi($tmpfname, $drm);
@@ -100,25 +104,73 @@ $csv = $export->export($mouvements);
 $t->is(count($csv), 9, "L'export genère 9 csv");
 
 $t->is(count($csv["01.DRMDEM"]), 1, "Une seul ligne pour les entrées");
-$t->is($csv["01.DRMDEM"][0], substr($periode,0,4).";".substr($periode,4,2).";".$viti->num_interne.";;0;40;30;20;10;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les entrées sont bien reportées");
+$t->is($csv["01.DRMDEM"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;40;30;20;10;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les entrées sont bien reportées");
 
 $t->is(count($csv["02.DRMDSA"]), 1, "Une seul ligne pour les sorties acquittées");
-$t->is($csv["02.DRMDSA"][0], substr($periode,0,4).";".substr($periode,4,2).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties acquittées sont bien reportées");
+$t->is($csv["02.DRMDSA"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties acquittées sont bien reportées");
 
 $t->is(count($csv["03.DRMDSS"]), 1, "Une seul ligne pour les sorties suspendues");
-$t->is($csv["03.DRMDSS"][0], substr($periode,0,4).";".substr($periode,4,2).";".$viti->num_interne.";;0;10;0;20;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties suspendues sont bien reportées");
+$t->is($csv["03.DRMDSS"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;10;0;20;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties suspendues sont bien reportées");
 
 $t->is(count($csv["04.DRMDSE"]), 1, "Une seul ligne pour les sorties exonérées");
-$t->is($csv["04.DRMDSE"][0], substr($periode,0,4).";".substr($periode,4,2).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties exonérées sont bien reportées");
+$t->is($csv["04.DRMDSE"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties exonérées sont bien reportées");
 
 $t->is(count($csv["05.DRMDSO"]), 1, "Une seul ligne pour les sorties autres");
-$t->is($csv["05.DRMDSO"][0], substr($periode,0,4).";".substr($periode,4,2).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties autres sont bien reportées");
+$t->is($csv["05.DRMDSO"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties autres sont bien reportées");
 
 $t->is(count($csv["06.DRMAX"]), 3, "3 lignes pour les exports");
 
 $t->is(count($csv["07.DRMCRD"]), 2, "2 lignes pour les CRDS");
 
 $t->is(count($csv["08.DRMENT"]), 1, "Une seul ligne pour le récap");
-$t->is($csv["08.DRMENT"][0], substr($periode,0,4).";".substr($periode,4,2).";".$viti->num_interne.";;0;\"\";0;0;0;221.39;44.28;265.67;30;265.67;40;30;20;10;0;0;0;0;10;0;20;0;0;0;0;0;0;0;0;0;".$dateFinMois.";".$date.";\"TELEDECLARATION\";\"\";\"\";;0.75;0.38;0;0", "Le recap est bien reporté");
+$t->is($csv["08.DRMENT"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;\"\";0;0;0;221.39;44.28;265.67;30;265.67;40;30;20;10;0;0;0;0;10;0;20;0;0;0;0;0;0;0;0;0;".$dateFinMois.";".$date.";\"TELEDECLARATION\";\"\";\"\";;0.75;0.38;0;0", "Le recap est bien reporté");
 
 $t->is(count($csv["09.ORIGINES"]), 13, "13 lignes de mouvements");
+
+$t->comment("Export d'une DRM à néant");
+
+$periode = DRMClient::getInstance()->getPeriodeSuivante($periode);
+$dateFinMois = new DateTime(DRMClient::getInstance()->buildDate($periode));
+$dateFinMois->modify('last day of this month');
+$dateFinMois = $dateFinMois->format('Ymd');
+
+$drm = DRMClient::getInstance()->createDoc($viti->identifiant, $periode);
+$drm->save();
+$drm->validate();
+$drm->save();
+
+$drm = DRMClient::getInstance()->find('DRM-'.$viti->identifiant.'-'.$periode);
+
+$mouvements = MouvementfactureFacturationView::getInstance()->getMouvementsAll(0);
+
+foreach($mouvements as $key => $mouvement) {
+    if($mouvement->id_doc == $drm->_id) {
+        continue;
+    }
+    unset($mouvements[$key]);
+}
+
+$export = new ExportMouvementsDRMDB2();
+$csv = $export->export($mouvements);
+
+$t->is(count($csv), 7, "L'export genère 7 csv");
+
+$t->is(count($csv["01.DRMDEM"]), 1, "Une seul ligne pour les entrées");
+$t->is($csv["01.DRMDEM"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les entrées sont vides");
+
+$t->is(count($csv["02.DRMDSA"]), 1, "Une seul ligne pour les sorties acquittées");
+$t->is($csv["02.DRMDSA"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties acquittées sont vides");
+
+$t->is(count($csv["03.DRMDSS"]), 1, "Une seul ligne pour les sorties suspendues");
+$t->is($csv["03.DRMDSS"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties suspendues sont vides");
+
+$t->is(count($csv["04.DRMDSE"]), 1, "Une seul ligne pour les sorties exonérées");
+$t->is($csv["04.DRMDSE"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties exonérées sont vides");
+
+$t->is(count($csv["05.DRMDSO"]), 1, "Une seul ligne pour les sorties autres");
+$t->is($csv["05.DRMDSO"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;0;0;0;0;0;0;0;0;".$date.";\"TELEDECLARATION\"", "Les sorties autres sont vides");
+
+$t->is(count($csv["08.DRMENT"]), 1, "Une seul ligne pour le récap");
+$t->is($csv["08.DRMENT"][0], substr($periode,0,4).";".(substr($periode,4,2)*1).";".$viti->num_interne.";;0;\"\";0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;".$dateFinMois.";".$date.";\"TELEDECLARATION\";\"\";\"\";;0;0;0;0", "Le recap est bien reporté");
+
+$t->is(count($csv["09.ORIGINES"]), 1, "1 ligne de mouvement");
